% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_p.R
\name{adjust_pmat}
\alias{adjust_pmat}
\title{adjust_p}
\usage{
adjust_pmat(pmat, p_adjust)
}
\arguments{
\item{pmat}{matrix of p-values to adjust}

\item{p_adjust}{character describing adjustment to make. See stats::p.adjust}
}
\description{
adjust_p
}
