% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{apa_matrix}
\alias{apa_matrix}
\title{apa matrix}
\usage{
apa_matrix(r_matrix, p_matrix, stars, round, remove_lead, triangle)
}
\arguments{
\item{r_matrix}{correlation coefficient matrix}

\item{p_matrix}{p-value matrix}

\item{stars}{a vector of pvalue stars}

\item{round}{How many digits to round to?}

\item{remove_lead}{a logical. Should leading zeros be removed?}

\item{triangle}{can select lower upper or NULL}

\item{show_ci}{bool. If TRUE, confidence interval will be displayed for non-partial pearson correlations.}
}
\description{
Creates an apa matrix
}
