% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{apa_table.corx}
\alias{apa_table.corx}
\title{apa_table.corx}
\usage{
apa_table.corx(corx, ...)
}
\arguments{
\item{corx}{corx object}

\item{...}{Other arguments to papaja::apa_table}
}
\description{
method for papaja::apa_table for corx objects
}
