% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{check_classes}
\alias{check_classes}
\title{check_classes}
\usage{
check_classes(data, ok_classes, stop_message, stop = TRUE)
}
\arguments{
\item{data}{the data object}

\item{ok_classes}{a vector of allowed classes}

\item{stop_message}{a character string provided to users if error triggers.}

\item{stop}{should the variable stop, or create a warning?}
}
\description{
check all classes are as expected
}
