% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{check_for_vec}
\alias{check_for_vec}
\title{check_for_vec}
\usage{
check_for_vec(names, sym, env)
}
\arguments{
\item{names}{names to check}

\item{sym}{names from call}

\item{env}{environment to search}
}
\description{
convert vectors of names to character vectors
}
