% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormat_list.R
\name{cormat_list}
\alias{cormat_list}
\title{cormat_list}
\usage{
cormat_list(data, x, y, z, method, p_adjust)
}
\arguments{
\item{data}{data.frame}

\item{x}{character vector, row names}

\item{y}{character vector, column names}

\item{z}{character vector, partial variable names}

\item{method}{string, passed to cor.test}

\item{p_adjust}{string, passed to p.adjust}
}
\description{
cormat_list
}
