% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{get_cor}
\alias{get_cor}
\title{get_cor}
\usage{
get_cor(data, x, y, method, partial)
}
\arguments{
\item{data}{data}

\item{x}{variable 1}

\item{y}{variable 2}

\item{method}{correlation method}

\item{partial}{control for anything?}
}
\description{
A flexible correlation function
}
