% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partial_matrix.R
\name{partial_matrix}
\alias{partial_matrix}
\title{partial_matrix}
\usage{
partial_matrix(data, x, y, method, partial)
}
\arguments{
\item{data}{the data object}

\item{x}{horizontal variables}

\item{y}{veritical variables}

\item{method}{the methos}

\item{partial}{variables to partial out}
}
\description{
Creates matricies of partial correlations including r, n, and p
}
