% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{partial_n_matrix}
\alias{partial_n_matrix}
\title{partial_n_matrix}
\usage{
partial_n_matrix(data, x, y, z)
}
\arguments{
\item{data}{data.frame or matrix}

\item{x}{rownames}

\item{y}{colnames}

\item{z}{partial variable vector}
}
\description{
Calculate complete observations for a crosstab + a third variable
}
