% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{plot.corx}
\alias{plot.corx}
\title{S3 class corx}
\usage{
\method{plot}{corx}(x, ...)
}
\arguments{
\item{x}{a corx object}

\item{...}{other arguments to ggcorrplot::ggcorrplot}
}
\description{
S3 class corx
}
