% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{plot_mds}
\usage{
plot_mds(corx, k = NULL, ...)
}
\arguments{
\item{corx}{the corx object, or a matrix of correlation coefficicents}

\item{k}{a numeric, the number of clusters. If set to "auto" will be equal to the number of principal components that explain more than 5\% of total variance.}

\item{...}{additional arguments passed to ggpubr::ggscatter}
}
\description{
plot the Classical multidimensional scaling of a corx object
}
