% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{rename_if_needed}
\alias{rename_if_needed}
\title{rename if needed}
\usage{
rename_if_needed(data, x)
}
\arguments{
\item{data}{data object}

\item{x}{a character vector. If named, columns will be renamed}
}
\description{
Renames columns
}
