% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{star_matrix}
\alias{star_matrix}
\title{star_matrix}
\usage{
star_matrix(m, stars)
}
\arguments{
\item{m}{matrix of p-values}

\item{stars}{a vector of p-value thresholds to replace with stars}
}
\description{
Replaces p-values with stars
}
