% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_clipboard.r
\name{to_clipboard}
\alias{to_clipboard}
\title{to_clipboard}
\usage{
to_clipboard(x, ...)
}
\arguments{
\item{x}{the corx object}

\item{...}{aditional arguments passed to clipr::write_clip}
}
\description{
Provides a way to send correlation tables to the clipboard so they can be pasted into excel.
}
