% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{corx}
\alias{corx}
\title{corx}
\usage{
corx(data, x = NULL, y = NULL, method = c("pearson", "spearman",
  "kendall"), partial = NULL, stars = c(0.05), round = 2,
  remove_lead = T, triangle = NULL, caption = NULL, note = NULL,
  ...)
}
\arguments{
\item{data}{A data.frame or matrix}

\item{x}{a vector of character names}

\item{y}{a vector of character names}

\item{method}{One of "pearson", "spearman", or "kendall"}

\item{partial}{a character vector of names. Used to perform partial correlation.}

\item{stars}{a numeric vector. Stars are added to summary if p-value is lower than each value.}

\item{round}{Number of digits in printing}

\item{remove_lead}{a bool. if true, leading zeros are removed in summaries}

\item{triangle}{one of "lower", "upper" or NULL \(default\)}

\item{caption}{table caption}

\item{note}{table note}
}
\description{
Create a correlation matrix
}
