% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{corx}
\alias{corx}
\title{corx}
\usage{
corx(data, x = NULL, y = NULL, method = c("pearson", "spearman",
  "kendall"), partial = NULL, stars = c(0.05), round = 2,
  remove_lead = T, triangle = NULL, caption = NULL, note = NULL,
  describe = F, grey_nonsig = T, ...)
}
\arguments{
\item{data}{A data.frame or matrix}

\item{x}{a vector of rownames. Defaults to all}

\item{y}{a vector of colnames. Defaults to all}

\item{method}{One of "pearson", "spearman", or "kendall"}

\item{partial}{a vector of colnames. Control variables to be used in partial correlations - defaults to NULL}

\item{stars}{a numeric vector. This argument defines cut-offs for p-value stars. This argument is vectorised}

\item{round}{Number of digits in printing}

\item{remove_lead}{a bool. if TRUE \(the default\), leading zeros are removed in summaries}

\item{triangle}{one of "lower", "upper" or NULL \(the default\)}

\item{caption}{table caption. Will be passed to plots}

\item{note}{table note}

\item{describe}{a list of functions with names or a logical. If functions are supplied to describe, a new column will be bound to the apa matrix for each item in the list. If TRUE is supplied, describe calculates means and standard deviations \(Missing values deleted listwise \)}

\item{grey_nonsig}{a bool. Should non-sigificant values be grey in output? This argument does nothing if describe is not set to FALSE}

\item{...}{additional arguments}
}
\value{
A list of class 'corx'.
}
\description{
Creates an object of class "corx". This function calculates correlation matricies. It stores effect sizes, p-values, the number of pairwise observations, and a formatted correlation matrix in a list. Partial correlations are calculated if the 'partial' argument is assigned values other than NULL. Methods are exported for the generic functions 'print', 'plot', summary, data.frame and 'coef'.
}
\examples{
cor_mat <- corx(mtcars, x = c(mpg,cyl,disp), y = c(wt,drat,disp,qsec), partial = wt, round = 2, stars = c(0.05), describe = list("mean" = function(x) mean(x,na.rm=TRUE)))
cor_mat
coef(cor_mat)
plot(cor_mat)
}
