% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corx.R
\name{par_matrix}
\alias{par_matrix}
\title{par_matrix}
\usage{
par_matrix(results, x, y, par)
}
\arguments{
\item{results}{results dataset}

\item{x}{one set of variables}

\item{y}{another set of variables}

\item{par}{the parameter to build with}
}
\description{
This function is used to construct final matricies
}
