% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mds.R
\name{plot_mds}
\alias{plot_mds}
\title{plot_mds}
\usage{
plot_mds(corx, k)
}
\arguments{
\item{corx}{the corx object, or a matrix of correlation coefficicents}

\item{k}{the number of clusters. If not provided, a prinipal components analysis is performed. k is set to the number of components which explain more than 5\% of variance}
}
\description{
plot the Classical multidimensional scaling of a corx object
}
