% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_clipboard.r
\name{to_clipboard}
\alias{to_clipboard}
\title{to_clipboard}
\usage{
to_clipboard(x, ...)
}
\arguments{
\item{x}{a corx object, matrix, or data.frame}

\item{...}{additional arguments passed to clipr::write_clip}
}
\description{
Sends a formatted corx table to the clipboard so that it can be pasted into excel.
}
