% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_table.R
\name{to_table}
\alias{to_table}
\title{to_table}
\usage{
to_table(corx, include_p = FALSE)
}
\arguments{
\item{corx}{a corx object}

\item{include_p}{logical. should p-values be included?}
}
\description{
Tabulate correlation matrices
}
